Allow specifying auth account name in the username
When authenticating with auth v1 as a user that has access to multiple
accounts, it's sometimes necessary to specify which account you want
to access.

This commit allows setting an account separator character, which is
used to split the username to find the account name, a bit like
keystone tenant config. The account name is then appended to the
authentication url.

So a username of "john:exampleltd" would generate an auth url of
https://auth.swift/v1/exampleltd

If no separator is specified, or no separator is found in the
username, the authentication url remains unmodified (which usually
means the default account is selected by the Swift server).